Add usableAdapter that will return the first usable (powered) adapter connected to your computer.
The rationale behind this is that we can have more than one adapter connected
to our computer, and the default one (usually the builtin one, if present) could
be powered off for energy saving or whatever reason. This method will return the
first adapter found that is ready to be used.