HACK: emit deviceFound in any case.
Because of the cache and the way bluez sends the information (it basically emits several times device found over the same devices, but the
information for the device could have been improved), the stable version is not ready to handle correctly this workflow.

With this new signal being emitted every time, bluedevil can take into account possible name changes and other property changes.

This signal will be removed for the next major release, which implements a correct way of handling the discovering process.

Conflicts:

	bluedevil/bluedeviladapter.cpp