HACK: if we have no default adapter check that we have adapter list and return the first one.
bluez sometimes reports that there is no default adapter, but there are adapters in the list
that the manager provides. In this case return the first adapter shown before returning that there
is no default adapter to be used. This slightly changes the behaviour of defaultDevice, but helps
to make life easier to libbluedevil users.