Added new api Device::registerDeviceAsync using pair technique
In Applications that need precision the creation of the device
need to be handler by theirself, so the registerDevice method
was added. this method is just like registerDevice async, using
the QDBusPendingCallWatcher so we don't have to use any kind of
threading.

A good example for these kind of application is the
bluedevil wizard.