Added Device::isReady which checks if it has been created
This is a little bit hackish but untul 2.0 we can't do anything better.
Basically we are checking if the device has been created in the bus and
use that to decide if it is "ready" or not.

this is the only way we have to know if a device actually works.